Use SCSSCacher to build theming css

+++ lib/private/Template/SCSSCacher.php
@@ -244,7 +244,9 @@ class SCSSCacher {
 	public function resetCache() {
 		foreach ($this->appData->getDirectoryListing() as $folder) {
 			foreach ($folder->getDirectoryListing() as $file) {
-				$file->delete();
+				if(substr($file->getName(), -3) === "css" || substr($file->getName(), -4) === "deps") {
+					$file->delete();
+				}
 			}
 		}
 	}
